Revert "Support organization icons and fix date content shift editing (#521)"
This reverts commit f9f47c965913719df6030f4ba47bf3e862b8530c.
