+++ apps/web/app/dashboard/caps/components/CapCard.tsx
@@ -31,14 +31,14 @@ interface Props extends PropsWithChildren {
     createdAt: Date;
     totalComments: number;
     totalReactions: number;
-    sharedOrganizations?: { id: string; name: string; iconUrl: string }[];
+    sharedOrganizations?: { id: string; name: string }[];
     ownerName: string | null;
     metadata?: VideoMetadata;
   };
   analytics: number;
   onDelete?: (videoId: string) => Promise<void>;
   userId?: string;
-  userOrganizations?: { id: string; name: string; iconUrl: string }[];
+  userOrganizations?: { id: string; name: string }[];
   sharedCapCard?: boolean;
   isSelected?: boolean;
   onSelectToggle?: () => void;
@@ -413,9 +413,9 @@ export const CapCard = ({
               )}
             </div>
             {renderSharedStatus()}
-            <div className="mb-1 h-[1.5rem]"> {/* Fixed height container */}
+            <div className="mb-1">
               {isDateEditing && !sharedCapCard ? (
-                <div className="flex items-center h-full">
+                <div className="flex items-center">
                   <input
                     type="text"
                     value={dateValue}
@@ -423,14 +423,14 @@ export const CapCard = ({
                     onBlur={handleDateBlur}
                     onKeyDown={handleDateKeyDown}
                     autoFocus
-                    className="text-sm w-full truncate text-gray-10 bg-transparent focus:outline-none h-full leading-[1.5rem]"
+                    className="text-sm truncate mt-2 leading-[1.25rem] text-gray-10 bg-transparent focus:outline-none"
                     placeholder="YYYY-MM-DD HH:mm:ss"
                   />
                 </div>
               ) : (
                 <Tooltip content={`Cap created at ${effectiveDate}`}>
                   <p
-                    className="text-sm truncate text-gray-10 cursor-pointer flex items-center h-full leading-[1.5rem]"
+                    className="text-sm truncate mt-2 leading-[1.25rem] text-gray-10 cursor-pointer flex items-center"
                     onClick={handleDateClick}
                   >
                     {showFullDate
