+++ apps/web/components/forms/NewOrganization.tsx
@@ -7,14 +7,9 @@ import {
   FormControl,
   FormField,
   Input,
-  CardDescription,
-  Label,
 } from "@cap/ui";
 import { useForm } from "react-hook-form";
 import { createOrganization } from "./server";
-import { useState } from "react";
-import { toast } from "sonner";
-import { FileInput } from "@/components/FileInput";
 
 export interface NewOrganizationProps {
   onOrganizationCreated: () => void;
@@ -35,9 +30,6 @@ export const NewOrganization: React.FC<NewOrganizationProps> = (props) => {
     },
   });
 
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-  const [isUploading, setIsUploading] = useState(false);
-
   return (
     <Form {...form}>
       <form
@@ -45,25 +37,12 @@ export const NewOrganization: React.FC<NewOrganizationProps> = (props) => {
         ref={props.formRef}
         onSubmit={form.handleSubmit(async (values) => {
           try {
-            setIsUploading(true);
             props.setCreateLoading?.(true);
-            
-            // Create FormData to send both the organization name and icon file
-            const formData = new FormData();
-            formData.append('name', values.name);
-            
-            // Add the icon file if one was selected
-            if (selectedFile) {
-              formData.append('icon', selectedFile);
-            }
-            
-            await createOrganization(formData);
+            await createOrganization(values);
             props.onOrganizationCreated();
           } catch (error) {
             console.error("Error creating organization:", error);
-            toast.error("Failed to create organization");
           } finally {
-            setIsUploading(false);
             props.setCreateLoading?.(false);
           }
         })}
@@ -85,23 +64,6 @@ export const NewOrganization: React.FC<NewOrganizationProps> = (props) => {
               </FormControl>
             )}
           />
-
-          <div className="space-y-1">
-            <Label htmlFor="icon">Organization Icon (Optional)</Label>
-            <CardDescription className="w-full max-w-[400px]">
-              Upload a custom logo or icon for your organization.
-            </CardDescription>
-          </div>
-
-          <div className="relative mt-2">
-            <FileInput
-              id="icon"
-              name="icon"
-              onChange={setSelectedFile}
-              disabled={isUploading}
-              isLoading={isUploading}
-            />
-          </div>
         </div>
       </form>
     </Form>
